Create subdirs even when pre-creating fileset with no files
E.g., for makedir workload the number of entries (files) in
the fileset is zero but there are leafdirs. If pre-creation
of subdirs is not performed then makedirs do NOT create any
directories because mkdir() fails with ENOENT.

In general, it is probably a good idea to create all subdirs
no matter how many files or leaf directories are in a fileset.

Reported by Abutalib Aghayev <agayev@gmail.com>.